requirements.txt
# Face Recognition Attendance System Requirements
# Core dependencies
streamlit>=1.28.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# Face Recognition
face-recognition>=1.3.0
dlib>=19.24.0

# Database
sqlalchemy>=2.0.0

# Data handling & Reports
pandas>=2.0.0
openpyxl>=3.1.0
reportlab>=4.0.0
xlsxwriter>=3.1.0

# Utilities
python-dateutil>=2.8.0
watchdog>=3.0.0

# Plotting
matplotlib>=3.7.0
plotly>=5.15.0
